backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
spacy
assemblyai
pytz
google-genai    
